#refactoring: (refactor) changed screenFrame structure, renamed "utils" file to "helpers"
