+++ packages/web/src/client-descriptors/deviceMemory.ts
@@ -1,4 +1,4 @@
-import { replaceNaN, toFloat } from '../utils/utils';
+import { replaceNaN, toFloat } from '../utils/helpers';
 
 export function getDeviceMemory(): number | null {
   // `navigator.deviceMemory` is a string containing a number in some unidentified cases
