+++ packages/web/src/client-descriptors/hardwareConcurrency.ts
@@ -1,4 +1,4 @@
-import { replaceNaN, toInt } from '../utils/utils';
+import { replaceNaN, toInt } from '../utils/helpers';
 
 export function getHardwareConcurrency(): number | null {
   return replaceNaN(toInt(navigator.hardwareConcurrency), null);
