+++ packages/web/src/client-descriptors/hdr.ts
@@ -1,4 +1,4 @@
-import { mediaQueryMatcher } from '../utils/utils';
+import { mediaQueryMatcher } from '../utils/helpers';
 
 export function isHDR(): boolean | null {
   const matchForDynamicRange = mediaQueryMatcher('dynamic-range') as (value: string) => boolean;
