+++ packages/web/src/client-descriptors/timezone.ts
@@ -1,4 +1,4 @@
-import { toFloat } from '../utils/utils';
+import { toFloat } from '../utils/helpers';
 
 export function getTimezone(): string {
   const DateTimeFormat = window.Intl?.DateTimeFormat;
