+++ packages/web/src/client-descriptors/touchSupport.ts
@@ -1,4 +1,4 @@
-import { toInt, hasProperty } from '../utils/utils';
+import { toInt, hasProperty } from '../utils/helpers';
 
 type TouchSupport = {
   maxTouchPoints: number;
