+++ packages/web/src/types/descriptors/index.ts
@@ -1,2 +1,3 @@
 export * from './webgl';
 export * from './webgpu';
+export * from './screenFrame';
