+++ packages/web/src/types/descriptors/screenFrame.ts
@@ -1,0 +1,9 @@
+/**
+ * The order matches the CSS side order: top, right, bottom, left
+ */
+export type FrameSize = [
+  number | null,
+  number | null,
+  number | null,
+  number | null,
+];
